короче я добавил хп врагам, экран победы и поражения, ну и ещё по мелочи. Кстати меню заработало
