+++ constants.py
@@ -1,0 +1,20 @@
+# Базовые константы
+WINDOW_SIZE = WINDOW_WIDTH, WINDOW_HEIGHT = 1920, 1080
+FPS = 30
+MAPS_DIR = 'maps'
+SPRITES_DIR = 'sprites'
+TILE_SIZE = 25
+MOVE_SPEED = 5
+ENEMY_EVENT_TYPE = 30
+ENEMY_DELAY = 200
+ENEMY_TRIGGER_SIZE = 25
+ENEMY_HP = 25
+PLAYER_HP = 25
+DAMAGE = 5
+username = 'Skerter'
+difficulty = 'Easy'
+kills = 0
+
+# Константы с цветами
+BLACK, WHITE, RED = (0, 0, 0), (255, 255, 255), (255, 0, 0)
+GREEN, BLUE, YELLOW = (0, 255, 0), (0, 0, 255), (255, 255, 0)
