+++ maps/map2.tmx
@@ -3,7 +3,7 @@
  <tileset firstgid="1" source="textures_hot_rooms.tsx"/>
  <layer id="1" name="Слой тайлов 1" width="50" height="35">
   <data encoding="base64" compression="zlib">
-   eJzNWNlywjAMdEuhTWAIR/r/v9pkxhq2Gl2WA+2DXnBs7+pYyexLKVfFPhfbG+v/xVacx8XeFBsWOxnrr7L3xeZqu8U+BJwej7uw79V2qT4fq3FMHo9Ljdst4TP03dQRB8JLWA6VE3LxeLT67Ap3XYX7Wk3La4rPCb5r5YE+iuaDdod3llWfiD3Do3XP2YgJrc3G+nqXlJceD89HZ8NHVky0/OD5jpgtnUEcnMfK4dvBSbFuqd2Inki1hL9zvV3XbnA+fbOrHPbF1ieKdTYXNf5a3FCnNFxDxX0oD43xNLY1rzCGke8y+papc4zn1nusWtqax9ZG/bOnp/0FD97315ocIQaZeGu6S3X/jLnqXHHTHE21aWGP9EvOg/QB9S7DSbsbZ7aIdkd7gZZXOC9lao90LTsj4jlSL+D90quPqeR6RibPNR7e3dE63/IdQvUXPSvSr6I8aA7yajGCj+o8Gl/soVP5rUM4hxzLQwv5+6fFz0MQn1fnqMtzxcjfTKNg9/KYq3rePy1vRyum1E8knDOsc7z0Hc5xkTyyLKNRyAF7C8cp4R1gXwaz158ib0fKd4wD12vpzYx4ozlt5ZLFP6K9PN8zPh2EfS3v7y2M+zqTF5LuPvs/Kz7fbmHafKW9JVuwaXFt7R9ZHsSF3v/Rd9zIvrXO7tXnL4MH9x/NVZG4SDqdna8yeUe1QP1Fm+Gj/uudbbGHtMzX2D+ysd5S31CDW+pH0t2M/3pqdmJnaf+pezx6MPT2UTTSMWn28PzA/4drNfJfzxn8LD4bRvf+AFEAHkE=
+   eJzNWNluAjEMTKFAFyqWY/n/X+1GisXUjY84AfXBL2STzPgYO+xSSlfBDqvtlPX/YhnnabUPwabVvpX1d9lmtaXYdrXPCk6Lx72y7912KT4/FuOYLB6XErdbwGfou3kQl4xlXzghF4tH5B6661q5b3R8KOcjPDYNuOg+6Q7rLG0dsUd4tO45KzGhtUVZz3fV8tLiYfnonNo0jOcAX+P5jpg1nUEcnEfm8DBwkha31K5HT2q1hL9zvc1rNzifvtkWDruk6xPFOpqLEn8pbqhTEq6p4N6np8ZYGtuaVxhDz3cRfYvUOcZz9B6tlkbzGG1cV96hoaNwY9/PNYm5FIm3pLtU96+Yq3J95RmJ5miqTQ271QtqPEgfUO8inKS7cWbzaLe3F0h5hfNSpPZI13pnxHOq9wLeL636mFOsZ0TyXOJh3e2t85HvEKo/71mefuXlQXOQVYsefFTn3vhiD53Tbx3COeSUnlrI3z8tfp6c+Kw6R11e0l9dpnu43dNzrup5/7S8HbWYPhScC6xzvPQdznGePNIsolHIAXsLx1nDO8G+COae9xyvK4wD1+vamxnxenNayyWNv0d7eb5HfDpV9rW8v0cY93UkL2q6++r/rEiPRp4vzVfSW7IFmxTX1v4R5UFc6P3vfccd2bfa2b36/KXw4P6jucoTl5pOR+erSN5RLVB/kWZ4r/96Z1vsIS3zNfaPaKxH6htqcEv91HQ34r+emp3ZWdJ/6haPHgy9fRSNdKw2e1h+4P/DtRr5r+cMfhafDb17fwCduR4J
   </data>
  </layer>
 </map>
